Fix: fix unsigned int

+++ arch/x86/x86_definitions.h
@@ -18,7 +18,7 @@
 
 #if (LINUX_VERSION_CODE < KERNEL_VERSION(2,6,16)) && defined (CONFIG_X86_32)
 #define native_read_cr4_safe() ({		\
-	unsigned int __dummy;			\
+	unsigned long __dummy;			\
 	__asm__("1:movl %%cr4, %0           \n"	\
 		"2:                         \n"	\
 		".section __ex_table,\"a\"  \n"	\
@@ -39,7 +39,7 @@
 #define native_read_cr2 read_cr2
 #else
 #define native_read_cr2() ({				\
-	unsigned int __dummy;				\
+	unsigned long __dummy;				\
 	__asm__ __volatile__(				\
 			     "movl %%cr2, %0\n\t"	\
 			     :"=r" (__dummy));		\
@@ -51,7 +51,7 @@
 #define native_read_cr3 read_cr3
 #else
 #define native_read_cr3() ({		\
-	unsigned int __dummy;		\
+	unsigned long __dummy;		\
 	__asm__ (			\
 		 "movl %%cr3, %0\n\t"	\
 		 :"=r" (__dummy));	\
